# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI Ticket",
        "version": "0.1.0"
    },
    "servers": [
        {
          "url": "https://santrixticket-1.1b7hlo69yoj4.eu-de.codeengine.appdomain.cloud"
        }
    ],   
    "paths": {
        "/create_ticket_mail/{request}": {
            "post": {
                "summary": "Send Email",
                "operationId": "send_email_create_ticket_mail__request__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/openapi.json": {
            "get": {
                "summary": "Custom Openapi Json",
                "operationId": "custom_openapi_json_openapi_json_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Response": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "success",
                    "message"
                ],
                "title": "Response"
            },
            "Ticket": {
                "properties": {
                    "mail_address": {
                        "type": "string",
                        "title": "Mail Address"
                    },
                    "hardware": {
                        "type": "string",
                        "title": "Hardware"
                    },
                    "hardware_brands": {
                        "type": "string",
                        "title": "Hardware Brands"
                    },
                    "tickettypID": {
                        "type": "integer",
                        "title": "Tickettypid"
                    },
                    "priority": {
                        "type": "string",
                        "title": "Priority"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "telefonnummer": {
                        "type": "string",
                        "title": "Telefonnummer"
                    },
                    "session_history": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Session History"
                    }
                },
                "type": "object",
                "required": [
                    "mail_address",
                    "hardware",
                    "hardware_brands",
                    "tickettypID",
                    "priority",
                    "username",
                    "telefonnummer",
                    "session_history"
                ],
                "title": "Ticket"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                                    "type": "string"
                                },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}